CPU-separation: move selftests
